return float of delim list instead of int

+++ mann/helper.py
@@ -18,7 +18,7 @@ def convert_str_to_int_array(string, delim=','):
     assert isinstance(string, str), "string parameter passed is not type str"
     assert isinstance(delim, str), "delim parameter passed is not type str"
     assert string != '', 'string parameter is empty'
-    return np.array([int(s) for s in string.strip().split(delim)])
+    return np.array([float(s) for s in string.strip().split(delim)])
 
 def convert_list_to_delim_str(list_to_convert, delim=','):
     """Return a string delimited by delim from a list
